Add license header.
